% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types_to_values.R
\name{apply_type_bool_and_error}
\alias{apply_type_bool_and_error}
\title{Within a list of TODO}
\usage{
apply_type_bool_and_error(values, types)
}
\arguments{
\item{values}{\code{list} of cell contents}

\item{types}{\code{character} or \code{list} of cell types ("s", "b", "str", or NA for numeric)}
}
\value{
Return a \code{list} of the values provided,  some values are turned into \code{logical} or \code{NA} (Excel error).
}
\description{
Within a list of TODO
}
\examples{
\dontrun{
url = "https://go.microsoft.com/fwlink/?LinkID=521962"
content <- read_xlsx_raw_content(url = url, sheet = 1)
cells_list  <- parse_sheet_xml_into_cells_list(content$sheet)
cells_list$values <- apply_type_bool_and_error(values = cells_list$values
                                             , types = cells_list$types )
}

}
