% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types_to_values.R
\name{apply_type_excel_dates}
\alias{apply_type_excel_dates}
\title{The 'style' value of a cell within a 'sheet.xml' need to be matched with the 'styles.xml' file : the value within the 'style' xml of a cell is expressing a position within the 'styles.xml'.
Once replaced, some of these styles are matched with a \code{date_numFmtId} parameters that is supposed - by default - to list the Excel-Date styles, i.e. styles from 14 to 22.}
\usage{
apply_type_excel_dates(
  values,
  styles,
  style_numFmtId,
  date_numFmtId = c(14, 15, 16, 17, 18, 19, 20, 21, 22)
)
}
\arguments{
\item{values}{\code{list} of cell contents}

\item{styles}{\code{list} of cell styles ("1", "2", "3", or NA)}

\item{style_numFmtId}{\code{numeric} - Ordered list of the styles within the \emph{styles.xml} file (numFmtId). They will be associated with some xml styles expressed within a cell of the sheet.xml.}

\item{date_numFmtId}{\code{numeric} - List of the styles id that are \code{Date}.}
}
\value{
Return a \code{list} of the values, some are turned into \code{Date} accordingly to their styles and the styles.xml file.
}
\description{
The 'style' value of a cell within a 'sheet.xml' need to be matched with the 'styles.xml' file : the value within the 'style' xml of a cell is expressing a position within the 'styles.xml'.
Once replaced, some of these styles are matched with a \code{date_numFmtId} parameters that is supposed - by default - to list the Excel-Date styles, i.e. styles from 14 to 22.
}
\examples{
\dontrun{
xmls <-  read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
xml_text <- xmls$sheet
cells_list  <- parse_sheet_xml_into_cells_list(xml_text)
cells_list$values <- apply_type_excel_dates(values = cells_list$values , styles = cells_list$styles
,  style_numFmtId = parse_styles_xml_sheet(xmls$styles))

}

}
