% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types_to_values.R
\name{apply_type_numeric}
\alias{apply_type_numeric}
\title{Convert Excel numeric cells from character to numeric}
\usage{
apply_type_numeric(values, types)
}
\arguments{
\item{values}{\code{list} of cell contents}

\item{types}{\code{character} vector of cell types ("s", "b", "str", or NA for numeric)}
}
\value{
Return a \code{list} of the values, numeric cells are converted to numeric and others unchanged
}
\description{
Convert Excel numeric cells from character to numeric
}
\examples{
\dontrun{
apply_type_numeric(c("42", "TRUE", "hello", "123"), c(NA, "b", "str", NA))
# Return the same return as : list(42, "TRUE", "hello", 123)

Other examples within the xlsxread pipeline :
xmls <-  read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
xml_text <- xmls$sheet
cells_list  <- parse_sheet_xml_into_cells_list(xml_text)
cells_list$values <- apply_type_bool_and_error(values = cells_list$values
, types = cells_list$types )

cells_list$values <- apply_type_numeric(values = cells_list$values
, types = cells_list$types)
}
}
