% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types_to_values.R
\name{apply_type_sharedstrings}
\alias{apply_type_sharedstrings}
\title{Convert character values accordingly to sharedstyles.xml}
\usage{
apply_type_sharedstrings(values, types, indexed_shared_values)
}
\arguments{
\item{values}{\code{list} of cell contents}

\item{types}{\code{character} vector of cell types ("s", "b", "str", or NA for numeric)}

\item{indexed_shared_values}{\code{character} vector of the sharedstrings.xml values}
}
\value{
Return a \code{list} of the values, some \code{character} values are converted accordingly to the sharedstyles.xml file.
}
\description{
Convert character values accordingly to sharedstyles.xml
}
\examples{
\dontrun{
xmls <-  read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
xml_text <- xmls$sheet
cells_list  <- parse_sheet_xml_into_cells_list(xml_text)
shared_values <- parse_sharedstrings( xmls$sharedstrings )

 cells_list$values <- apply_type_sharedstrings(values = cells_list$values
, types = cells_list$types
, indexed_shared_values = shared_values)

cells_list$values <- apply_type_numeric(values = cells_list$values, types = cells_list$types)
}

}
