% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types_to_values.R
\name{apply_types_to_sheet_xml_cells_list}
\alias{apply_types_to_sheet_xml_cells_list}
\title{Apply types to parsed list of results}
\usage{
apply_types_to_sheet_xml_cells_list(
  values,
  types,
  styles,
  indexed_shared_values,
  styles_numFmtId,
  ...
)
}
\arguments{
\item{values}{\code{list} - Cells contents}

\item{types}{\code{character} or \code{list} - Cells types ("s", "b", "str", or NA for numeric)}

\item{styles}{\code{list} of cell styles ("1", "2", "3", or NA)}

\item{indexed_shared_values}{\code{character} - The sharedstrings.xml values, e.g., returned by \code{parse_sharedstrings()}}

\item{styles_numFmtId}{\code{numeric} - Ordered list of the styles within the \emph{styles.xml} file (numFmtId). They will be associated with some xml styles expressed within a cell of the sheet.xml.}

\item{...}{\itemize{
\item Other parameters passed to \code{apply_type_excel_dates()}, i.e. tweak the values considered as Excel-Date styles with the parameter \code{date_numFmtId}
}}
}
\value{
Return the \code{values} provided as a \code{list}, with some proper \code{logical} and \code{NA} (Excel error) values.
}
\description{
Apply types to parsed list of results
}
\examples{
\dontrun{
url = "https://go.microsoft.com/fwlink/?LinkID=521962"
content <- read_xlsx_raw_content(url = url, sheet = 1)
cells_list  <- parse_sheet_xml_into_cells_list(content$sheet)
shared_values <- parse_sharedstrings( content$sharedstrings )
styles_numFmtId <- parse_styles_xml_sheet(content$styles)

final_content <- apply_types_to_sheet_xml_cells_list(values = cells_list$values
, types = cells_list$types
, styles = cells_list$styles
, indexed_shared_values = shared_values
, styles_numFmtId = styles_numFmtId)
}

}
