% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_df_from_cells_list.R
\name{build_df_from_cells_list}
\alias{build_df_from_cells_list}
\title{Build a data.frame from a cells list, preserving types (Date, numeric, character)}
\usage{
build_df_from_cells_list(cells_list, remove_first_row_as_header = TRUE)
}
\arguments{
\item{cells_list}{A list containing at least:
\itemize{
\item values: list of cell contents (each element can be numeric, character, Date, etc.)
\item row: vector or list of row numbers corresponding to each value
\item col: vector or list of column letters corresponding to each value
}}

\item{remove_first_row_as_header}{Logical, if TRUE the first row is used as column names}
}
\value{
A \code{data.frame} with optional headers
}
\description{
Build a data.frame from a cells list, preserving types (Date, numeric, character)
}
\examples{
\dontrun{

url = "https://go.microsoft.com/fwlink/?LinkID=521962"
content <- read_xlsx_raw_content(url = url, sheet = 1)
cells_list  <- parse_sheet_xml_into_cells_list(content$sheet)
shared_values <- parse_sharedstrings( content$sharedstrings )
styles_numFmtId <- parse_styles_xml_sheet(content$styles)

cells_list$values <- apply_types_to_sheet_xml_cells_list(values = cells_list$values
, types = cells_list$types
, indexed_shared_values = shared_values
, styles_numFmtId = styles_numFmtId)

datas <- build_df_from_cells_list(cells_list, remove_first_row_as_header = TRUE)
str(datas)
}

}
