% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inlineStr_text.R
\name{extract_inlineStr_text}
\alias{extract_inlineStr_text}
\title{Extract text from inlineStr XML cell content}
\usage{
extract_inlineStr_text(raw_content)
}
\arguments{
\item{raw_content}{character vector containing inlineStr XML}
}
\value{
character vector of extracted text (NA if no \if{html}{\out{<t>}} node)
}
\description{
Extract text from inlineStr XML cell content
}
\details{
The inlineStr text is supposed to be between \if{html}{\out{<t>}} ... \if{html}{\out{</t>}} but other types of inline content are not handled yet
}
\keyword{internal}
