% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_cells_values.R
\name{parse_raw_cells_values}
\alias{parse_raw_cells_values}
\title{Get all the cells values (raws) before to turn it into a matrix}
\usage{
parse_raw_cells_values(sheet_text = NULL)
}
\arguments{
\item{sheet_text}{\code{character} - The complete raw sheet text (e.g., from 'sheet1.xls')}
}
\value{
Return a \code{data.frame} with
\itemize{
\item \code{raw_content} - The raw content
\item \code{raw_value} - The value within the sheet .xml (index or real numeric value)
\item \code{excel_index} - The excel_index given in the .xml (start from 0 for indexing the first entry of the sharedstrings .xml file)
\item \code{types} - The types of value with
\itemize{
\item \code{NA} for numeric,
\item "s" for shared string,
\item "inlineStr" for inline text value,
\item "b" for a logical boolean,
\item "str" for a formula result,
\item "e" for an Excel error,
\item "d" for ISO date
}
}
}
\description{
Get all the cells values (raws) before to turn it into a matrix
}
\examples{
\dontrun{
excel_df <- parse_raw_cells_values(
sheet_text = read_xlsx_raw_content(
url = "https://go.microsoft.com/fwlink/?LinkID=521962")$sheet )

str(excel_df)
}

}
