% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sheet_xml.R
\name{parse_sheet_xml_into_cells_list}
\alias{parse_sheet_xml_into_cells_list}
\title{Parse the raw - text - content of sheet.xml into a list of cells}
\usage{
parse_sheet_xml_into_cells_list(xml_text)
}
\arguments{
\item{xml_text}{\code{character} - The text of a sheet.xml from an .xlsx file}
}
\value{
Return a \code{list} with one entry per row.
Each entry is a \code{list} with a string representing a cell : all returned values are \code{character} excepted the 'col' entry (\code{numeric}).
For exemple within the first entry, returned values are :
\itemize{
\item entry 1
\item \code{ref} - \code{character} : Excel-style complete reference of the cell, e.g., 'A1' for the first cell
\item \code{row} - \code{numeric} : Excel-style reference of the row, e.g., '1'
\item \code{col} - \code{character} :  Excel-style reference of the col, e.g., 'A'
\item \code{type} - \code{character} :  xml-style type of cell, e.g., e.g., 's'
\item \code{values} - \code{character} raw value expressed in the .xml, e.g., '100'
}
}
\description{
Parse the raw - text - content of sheet.xml into a list of cells
}
\examples{
\dontrun{
xmls <-  read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
xml_text <- xmls$sheet
cells_list  <- parse_sheet_xml_into_cells_list(xml_text)
}

}
