% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_styles.xml_sheet.R
\name{parse_styles.xml_sheet}
\alias{parse_styles.xml_sheet}
\title{Parse styles.xls sheet from an .xlsx Workbook into a data.frame}
\usage{
parse_styles.xml_sheet(styles_raw)
}
\arguments{
\item{styles_raw}{\code{character} - The style raw text content as returned by \code{read_xlsx_sheet_raw_content()}}
}
\value{
\itemize{
\item Return a \code{data.frame} with 2 columns : \code{numFmtId} and \code{styles_index}
}
}
\description{
Parse styles.xls sheet from an .xlsx Workbook into a data.frame
}
\examples{
 parse_styles.xml_sheet(styles_raw = read_xlsx_sheet_raw_content(
 url = "https://go.microsoft.com/fwlink/?LinkID=521962")$styles)
}
