% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_styles.xml_sheet.R
\name{parse_styles_xml_sheet}
\alias{parse_styles_xml_sheet}
\title{Parse styles.xls sheet from an .xlsx Workbook into a data.frame}
\usage{
parse_styles_xml_sheet(styles_raw)
}
\arguments{
\item{styles_raw}{\code{character} - The style raw text content as returned by \code{read_xlsx_sheet_raw_content()}}
}
\value{
\itemize{
\item Return a \code{character} vector of \code{numFmtId}.
}
}
\description{
Parse styles.xls sheet from an .xlsx Workbook into a data.frame
}
\details{
The .xml convention begin the index of the \code{numFmtId} from '0', e.g., the first numFmtId is associated with the sheet.xml values '0'.
Please note that R natural index begin with 1.
}
\examples{
\dontrun{
 parse_styles_xml_sheet(styles_raw = read_xlsx_raw_content(
 url = "https://go.microsoft.com/fwlink/?LinkID=521962")$styles)
 }
}
