% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_df_from_xml_xlsx.R
\name{pivot_df_from_xml_xlsx}
\alias{pivot_df_from_xml_xlsx}
\title{Build a data.frame from a df computed with replace_values_with_types(), i.e. according to a raw_value and a excel_index column of the data.frame provided by the user.}
\usage{
pivot_df_from_xml_xlsx(index_df)
}
\arguments{
\item{index_df}{\code{data.frame} - The \code{data.frame} computed by \code{replace_values_with_types()}. Must contain the following columns : excel_index and raw_value}
}
\value{
Return a wild \code{data.frame}, similar to Excel
}
\description{
Build a data.frame from a df computed with replace_values_with_types(), i.e. according to a raw_value and a excel_index column of the data.frame provided by the user.
}
\examples{
\dontrun{
content <- read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")

index_df <- replace_values_with_types(
parse_raw_cells_values(content$sheet)
, parse_sharedstrings(text = content$sharedstrings))

xlsx_data <- pivot_df_from_xml_xlsx(index_df)

str(xlsx_data)
}
}
