% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx_raw_content.R
\name{read_xlsx_raw_content}
\alias{read_xlsx_raw_content}
\title{Read .xml files from - unzipped - .xlsx and return raw content}
\usage{
read_xlsx_raw_content(file = NULL, sheet = 1, url = NULL, ...)
}
\arguments{
\item{file}{\code{character} - Path to a .xlsx file to read}

\item{sheet}{\code{numeric} - Number of the sheet to extract, default = 1}

\item{url}{\code{character} - url to a .xlsx file to read (link with no waiting page only)}

\item{...}{\itemize{
\item Parameters passed to \code{utils::download.file()}
}}
}
\value{
\code{list} - Return the raw .xml content (unparsed) for several entries within the "xl/" subfolder :
\itemize{
\item \code{sheet} - raw data for one of the sheet, default is the first sheet (i.e. "xl/sheet1.xml"). This entry already offers the numeric values but the others values are a sharedStrings-index numeric value (the next \code{'sharedstrings'} entry of the \code{list} of results is the mapping file)
\item \code{sharedstrings} - sharedStrings.xml raw data (mapping file for characters and special values)
\item \code{workbook} - workbook.xml raw data (overall informations)
\item \code{styles} - styles.xml raw data
}
}
\description{
Read .xml files from - unzipped - .xlsx and return raw content
}
\examples{
content <-  read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
str(content)
}
