% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxlsx.R
\name{readxlsx}
\alias{readxlsx}
\title{Read a sheet from an .xlsx file (local filepath or url)}
\usage{
readxlsx(filepath = NULL, url = NULL, sheet = 1, row1_si_colnames = F, ...)
}
\arguments{
\item{filepath}{\code{character} - Local path to the .xlsx file to read}

\item{url}{\code{character} - url to an .xlsx file to download and read, downloading is processed with \code{utils::download.file()}}

\item{sheet}{\code{numeric} - Number of the sheet to extract, default = 1}

\item{row1_si_colnames}{\code{logical} - Use the first row values as colnames if \code{TRUE}. Default is \code{FALSE}.}

\item{...}{\itemize{
\item Parameters passed to \code{utils::download.file()}
}}
}
\description{
Read a single sheet from a .xlsx and guess the correct style (e.g., \code{Date}, \code{numeric} values)
}
\examples{
datas <- readxlsx(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
str(datas)
}
