% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values_with_types.R
\name{replace_values_with_types}
\alias{replace_values_with_types}
\title{Replace values from a 'raw_value' column of a df, according to another dictionnary df}
\usage{
replace_values_with_types(cells_df, sharedstrings_df)
}
\arguments{
\item{cells_df}{\code{data.frame} - The raw content data.frame provided by parse_raw_cells_values(). Must contain 'types', 'raw_content' and 'raw_value' columns.}

\item{sharedstrings_df}{\code{data.frame} - The content data.frame provided by parse_sharedstrings(). Must contain 'value' and 'index_excel' columns}
}
\value{
\itemize{
\item Return a \code{data.frame} similar to the one passed by the user (\code{cells_df} parameter), but the column 'raw_content' is tweaked several times (see details), e.g., accordingly to the sharedstrings \code{data.frame} (\code{sharedstrings_df} parameter)
}
}
\description{
Replace values according to the sharedstrings.xml file (e.g., character value of a cell are in this file) and transform other values into dates according to the styles.xml
}
\details{
The cells_df is processed accordingly to the 'types' of values :
\itemize{
\item types == 'inlineStr' - \code{character} inline values are extracted (i.e. text within '\if{html}{\out{<t>}}' and '\if{html}{\out{</t>}}'),
\item types == "s" - \code{character} values are permuted accordingly to the \code{sharedstrings_df} data.frame
\item types == "b" - \code{logical} values are turned into logical \code{TRUE} / \code{FALSE} / \code{NA}
\item types == "n" - \code{numeric} values are turned into numeric
}

Note that the date have to be post-processed accordingly to the styles.xml and are treated as numeric values by this function.

The XML entities such as '&' '<' are not translated
}
\examples{
\dontrun{
content <- read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")

cells_df = parse_raw_cells_values(content$sheet)
sharedstrings_df = parse_sharedstrings(text = content$sharedstrings)

index_df <- replace_values_with_types(cells_df, sharedstrings_df)

str(index_df)
}
}
