% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppress_balises}
\alias{suppress_balises}
\title{Remove \if{html}{\out{<t>}} tags from inlineStr matches}
\usage{
suppress_balises(x, balis = "t")
}
\arguments{
\item{x}{\code{character} - vector of strings containing a \code{balis} such as \if{html}{\out{<t>}}...\if{html}{\out{</t>}}}

\item{balis}{\itemize{
\item \code{character} - Text to use for considering a match such as \if{html}{\out{<text>}} ... \if{html}{\out{</text>}}, default = "t"
}}
}
\value{
character vector with only inner text
}
\description{
Remove \if{html}{\out{<t>}} tags from inlineStr matches
}
\examples{
\dontrun{
x <- c("<t>Hello</t>"
,"<t xml:space='preserve'>World</t>","<t>Hel</t><t>lo</t>")

suppress_balises(x, balis = "t")
}

}
