% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_xlsx}
\alias{unzip_xlsx}
\title{Unzip a .xlsx file into a bunch of .xml}
\usage{
unzip_xlsx(file = NULL, url = NULL, ...)
}
\arguments{
\item{file}{\code{character} - Path to a .xlsx file to unzip}

\item{url}{\code{character} - url to an .xlsx file to download and unzip, downloading is processed with \code{utils::download.file()}}

\item{...}{\itemize{
\item Parameters passed to \code{utils::download.file()}
}}
}
\value{
\code{character} - Return a temporary directory path : the - main - directory path of the bunch of .xml files produced
}
\description{
Unzip a .xlsx file into a bunch of .xml
}
\details{
When an url is provided, a binary download mode is used, i.e. utils::download.file(mode = "wb").
}
\seealso{
\code{\link[utils:download.file]{utils::download.file()}}
}
