% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_extract}
\alias{xml_extract}
\title{Utility function : don't repeat regmatches(text, grepexpr) ad nauseam}
\usage{
xml_extract(text, pattern)
}
\arguments{
\item{text}{\code{character} - Text to process with regmatches(text, gregexpr( pattern , text ) )}

\item{pattern}{\code{character} - pattern to process}
}
\description{
Utility function : don't repeat regmatches(text, grepexpr) ad nauseam
}
