% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sharedstrings.R
\name{parse_sharedstrings}
\alias{parse_sharedstrings}
\title{Parse the sharedstrings.xml file}
\usage{
parse_sharedstrings(text)
}
\arguments{
\item{text}{\code{character} - The raw text content of a sharedstrings.xml file}
}
\value{
Return a \code{data.frame} with 3 columns :
\itemize{
\item \code{index_excel} for the Excel '0-first' index (line-number - 1)
\item \code{index_R} for the '1-first' index, as the main line numbering method within R
\item \code{value} for the raw value
}
}
\description{
sharedstrings.xml  is a dictionnary with the values within the .xlsx sheet
}
\examples{
\dontrun{
content <- read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")
parse_sharedstrings(text = content$sharedstrings)
}

}
