% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxlsx.R
\name{readxlsx}
\alias{readxlsx}
\title{Read a sheet from an .xlsx file (local filepath or url)}
\usage{
readxlsx(
  filepath = NULL,
  url = NULL,
  sheet = 1,
  remove_first_row_as_header = F,
  skip_rows = NULL,
  ...
)
}
\arguments{
\item{filepath}{\code{character} - Local path to the .xlsx file to read}

\item{url}{\code{character} - url to an .xlsx file to download and read, downloading is processed with \code{utils::download.file()}}

\item{sheet}{\code{numeric} - Number of the sheet to extract, default = 1}

\item{remove_first_row_as_header}{\code{logical} - Use the first row values as colnames if \code{TRUE}. Default is \code{FALSE}}

\item{skip_rows}{\code{numeric} - Skip first rows of the datas, \emph{before} to consider the first row of this filtered dataset as a header (if \code{remove_first_row_as_header} is \code{TRUE})}

\item{...}{\itemize{
\item Parameters passed to \code{utils::download.file()} if needed for the download (only used when the user provide the \code{url} parameter)
}}
}
\description{
Read a single sheet from a .xlsx and guess the correct style (e.g., \code{Date}, \code{numeric} values)
}
\examples{
datas <- readxlsx(url = "https://go.microsoft.com/fwlink/?LinkID=521962"
                ,  remove_first_row_as_header = TRUE)
str(datas)
}
