% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values_with_types.R
\name{replace_values_with_types}
\alias{replace_values_with_types}
\title{Permute values in a 'raw_value' column of a df, according to another dictionnary df}
\usage{
replace_values_with_types(raw_cells_df, sharedstrings_df, styles_df)
}
\arguments{
\item{raw_cells_df}{\code{data.frame} - The raw content data.frame provided by parse_raw_cells_values(). Must contain 'types', 'raw_content' and 'raw_value' columns.}

\item{sharedstrings_df}{\code{data.frame} - The content data.frame provided by parse_sharedstrings(). Must contain 'value' and 'index_excel' columns}

\item{styles_df}{\code{data.frame} - The data.frame provided by parse_styles.xml_sheet(). Must contain 'styles_index' and 'numFmtId' columns}
}
\value{
\itemize{
\item Return a \code{data.frame} similar to the one passed by the user (\code{raw_cells_df} parameter), but the column 'raw_content' contain the 'good' values, according to the sharedstyle \code{data.frame} (\code{sharedstrings_df} parameter)
}
}
\description{
Permute values according to the sharedstrings.xml file (e.g., character value of a cell are in this file) and transform other values into dates according to the styles.xml
}
\details{
XML entities such as '&' '<' are not transleted and returned as they are '&amp' : the user have to post-process with a real XML parser
}
\examples{
\dontrun{
content <- read_xlsx_raw_content(url = "https://go.microsoft.com/fwlink/?LinkID=521962")

cells_df = parse_raw_cells_values(content$sheet)
shstrings_df = parse_sharedstrings(text = content$sharedstrings)
styles_df = parse_styles.xml_sheet(content$styles))

index_df <- replace_values_with_types(cells_df, shstrings_df, styles_df)

str(index_df)
}
}
